requirements.txt
asgiref==3.5.2
async-timeout==4.0.2
certifi==2022.6.15
charset-normalizer==2.1.0
Deprecated==1.2.13
Django==4.0.6
django-environ==0.9.0
django-redis==5.2.0
djangorestframework==3.13.1
idna==3.3
packaging==21.3
Pillow==9.2.0
plyer==2.0.0
psycopg2==2.9.3
pymongo==4.2.0
pyparsing==3.0.9
pytz==2022.1
razorpay==1.3.0
redis==4.3.4
requests==2.28.1
sqlparse==0.4.2
tzdata==2022.1
urllib3==1.26.11
wrapt==1.14.1
